chore: update package main and test script
