+++ index.html
@@ -89,7 +89,7 @@
     <h2>1. People</h2>
     <div id="people-section">
       <input type="text" id="person-name" placeholder="Name" />
-      <button id="save-people" onclick="addPerson()">Add Person</button>
+      <button id="save-people">Add Person</button>
       <ul id="people-list"></ul>
     </div>
 
@@ -121,7 +121,7 @@
         "
         placeholder="Current state JSON"
       />
-      <button onclick="downloadJson()">Download JSON File</button>
+      <button id="download-json">Download JSON File</button>
     </div>
     <textarea
       id="state-json-input"
@@ -130,16 +130,13 @@
       placeholder="State JSON"
     ></textarea>
     <br />
-    <button onclick="loadStateFromJson()">Load from JSON</button>
-    <button onclick="document.getElementById('state-json-file').click()">
-      Load from JSON File
-    </button>
+    <button id="load-json">Load from JSON</button>
+    <button id="load-json-file">Load from JSON File</button>
     <input
       type="file"
       id="state-json-file"
       accept="application/json"
       style="display: none"
-      onchange="loadStateFromJsonFile(this.files[0])"
     />
 
     <h2>6. Share</h2>
@@ -159,6 +156,6 @@
     </div>
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/lz-string/1.5.0/lz-string.min.js"></script>
-    <script src="scripts.js"></script>
+    <script type="module" src="main.js"></script>
   </body>
 </html>
