+++ package.json
@@ -2,15 +2,15 @@
   "name": "cost-splits.github.io",
   "version": "1.0.0",
   "description": "",
-  "main": "scripts.js",
+  "main": "main.js",
   "scripts": {
-    "test": "jest",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
     "lint": "eslint ."
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "commonjs",
+  "type": "module",
   "prettier": {
     "printWidth": 80,
     "proseWrap": "always"
